requirements.txt
django~=4.2.5
djangorestframework~=3.14.0
python-dotenv~=1.0.0
gunicorn~=21.2.0
psycopg2-binary~=2.9.7
pre_commit~=3.4.0
djangorestframework-simplejwt~=5.3.0
drf-spectacular~=0.26.5
django-filter~=23.3
black~=23.9.1
